add test for master and puppet keywords
Related to [#108](https://github.com/Scony/godot-gdscript-toolkit/issues/108).
This test will fail until `puppet` is added to the grammar.

This is my first PR here, please let me know if I've done this correctly, or if I need to make any changes. I'm not familiar with how the parser/grammar works yet, but I figured at least we can avoid regression, as well as a not-so-subtle reminder that this issue should be fixed, with this test.